Render whitelist of glyphs with emoji font as the first priority, fixes #1267
Define the range of unicode symbols that are emojis according to the standard 5.0
Use `emoji` font as a first priority for these glyphs.
Add more emoji fonts to the list (possibly makes emojis nice without fontconfig adjustments)

This was already partially attempted to be done in #1040, but due to
not defining the whitelist it was reverted and replaced with #1057.