fix: Improve detection of missing members when an event is received
Currently, if we receive a `join` event, we will check that the sender is already in the conversation participants. If not we will manually add it. But this step is also done by actually parsing the `join` event.

Thus we can filter `join` events that are sent by the actual user joining the conversation.