fix: Decorrelate AV Preferences' mediaStream from the one used in a call
Previously, mediaStream were shared accross the preferences page and a potential current call. Which lead to hard-to-find mediaStream leaks (because it's hard to know if you need to release the tracks of the media stream from the call and/or the preference page).

Now, we have discinct workflow for both the page and a call. Which means they are both responsible for cleaning their own mediaStream, but never the media stream of someone else. It makes the dataflow easier to understand and to scale in the future