runfix: Avoid race condition when fetching users (#6930)
At app load time, we update the conversations (fetch the users and make sure everything is there). But since we did not wait for this to be over when loading the app, we would also load those users again when displaying the conversation. Resulting in 2 requests being sent in parallel and 2 user entities being created for the same user