chore(ci): Add job to update image tag in corresponding Helm chart (#11158)
This change-set is meant to keep the webapp Helm chart in wire-server
up-to-date. To accomplish that, it adds two new blocks of steps (one
of which actually incorporates a new job).

(1) following the naming scheme in ./bin/push_docker.js:59, a first
step assembles the image tag name (not the release tag) - and so the
naming clash begins. A second one writes the information to an 'artifact'.
This way, its possible to access the information from another job.

(2) this new job reads the image tag from the artifact and opens a
new PR over at wire-server. This PR is supposed to only update the
webapp chart - more precisely it only changes the value of 'image.tag'
in wire-server/charts/webapp/values.yaml

Finally, the artifact is removed by the general clean up job 'cleanup_artifacts'.

Furthermore, the change-set simplifies cleanup with the help of setting
'failOnError', which wont cause the workflow to fail, if a certain
artifact does not exist.