chore(wire-server): Include changelog entry
We recently added a changelog process on wire-server. This adds a file on the automatic PRs with a changelog description that is then picked up with the rest of our current process. This is for the helm charts of the webapp, also currently located inside the wire-server repo. These are used as the defaults for on-premise installations.