fix: Target lastRead and Countly messages
Since the migration to the core, we provided no users to target the message for lastRead and Countly payloads. This meant that the core was trying to find which users to send to and ended up downloading the full user list and generating payload for those ... Including the self device.

This fixes it by passing an explicit list of users to send the message to from the webapp