fix: Detect deleted clients when sending a message (#14755)
There could be situations where we know of clients, but do not have a session with them. Which means that, when we try to encrypt for those clients, we will try to get the prekeys from them to which the backend will say that there are no prekeys.
Previously we would just filter them out and ignore them. Now the core warns the webapp that those clients have been deleted and that the webapp should wipe them out from the DB