test: add new fixtures to create users and pages (#19767)
In a lot of cases we need to create multiple pages to simulate multiple devices / users within one test. The two new added fixtures createPage() and createUser() allow us to do that in a save way, ensuring tests are isolated and created contexts / users are correctly cleaned up. (Even if the test fails in the middle)
The fixtures also allow us to generalize some more advanced use cases through the API. For example preventing the Telemetry modal from appearing after every login, reducing the time a test takes by ~5s per login. The design of the fixtures is extensible and more options may be added in the future if they really provide benefit to a lot of tests.
I started by refactoring the login tests, using them as a POC. The ultimate goal would be to refactor all tests creating a uniform codebase but I'd like to avoid doing so within one PR, so please see this as the first step of an incremental refactoring.